Updated game scene, corrected errors
