+++ src/scenes/Title.js
@@ -12,7 +12,7 @@ export class Title extends Phaser.Scene {
     this.titleText.setOrigin(0.5);
 
     // create the Play game button
-    this.startGameButton = new UiButton(this, this.scale.width / 2, this.scale.height * 0.65, 'button1', 'button2', 'Start', this.startScene.bind(this, 'Game'));
+    this.startGameButton = new UiButton(this, this.scale.width / 2, this.scale.height * 0.65, 'button1', 'button2', 'Start', this.startScene.bind(this, 'CharacterSelection'));
     }
 
     startScene(targetScene) {
